MAR-2099 combined sections in Main.vue and combined data in caseClutch.js
