--- client/data/clutch.js
@@ -1,122 +1,0 @@
-export const reviewsStatistics = [
-  {
-    numberText: '1',
-    number: 1,
-    name: 'million',
-    numberColor: '#043AB4',
-    description: 'monthly buyers',
-    icon: {
-      name: 'account',
-      width: '30px',
-      height: '30px',
-      folder: 'clutch',
-      type: 'svg',
-    },
-  },
-  {
-    numberText: '225,800',
-    number: 225800,
-    name: '+',
-    numberColor: '#043AB4',
-    description: 'companies',
-    icon: {
-      name: 'analytics',
-      width: '30px',
-      height: '30px',
-      folder: 'clutch',
-      type: 'svg',
-    },
-  },
-  {
-    numberText: '104,300',
-    number: 104300,
-    name: '+',
-    numberColor: '#043AB4',
-    description: 'client reviewes',
-    icon: {
-      name: 'bookmark',
-      width: '30px',
-      height: '30px',
-      folder: 'clutch',
-      type: 'svg',
-    },
-  },
-]
-
-export const clutchAwards = [
-  {
-    description: 'Deloitte’s Technology Fast 500',
-    img: {
-      width: 120,
-      height: 120,
-      name: 'review-technologyFast',
-      type: 'webp',
-    },
-  },
-  {
-    description: 'LinkedIn Top 50 Startups in the U.S.',
-    img: {
-      width: 140,
-      height: 120,
-      name: 'review-linkedIn',
-      type: 'webp',
-    },
-  },
-  {
-    description: 'Inc 500 America’s Fastest-Growing Private Companies',
-    img: {
-      width: 120,
-      height: 120,
-      name: 'inc-500',
-      type: 'webp',
-    },
-  },
-]
-
-export const speedStatistics = [
-  {
-    numberText: '1',
-    number: 10,
-    name: 'second',
-    numberColor: '#ffffff',
-    headerText: 'First Contentful Paint',
-    headerTextColor: '#A0A0A1',
-  },
-  {
-    numberText: '1',
-    number: 10,
-    name: 'second',
-    numberColor: '#ffffff',
-    headerText: 'Speed Index',
-    headerTextColor: '#A0A0A1',
-  },
-  {
-    numberText: '1',
-    number: 10,
-    name: 'second',
-    numberColor: '#ffffff',
-    headerText: 'Last Painted Hero',
-    headerTextColor: '#A0A0A1',
-  },
-]
-
-export const dataRep = [
-  {
-    text: 'The service we developed improved the registration experience with the autofill of location-based inputs.',
-    background: '#59C6E9',
-  },
-  {
-    text: 'Location-based taxonomy enables Clutch.co users to find vendors from a given point within a given radius.',
-    background: '#1B85CE',
-  },
-  {
-    text: 'A user-friendly interface allows managing and editing locations.',
-    background: '#043AB4',
-  },
-]
-
-export const idDataRep = [
-  'data-item-0',
-  'data-item-1',
-  'data-item-2',
-]
